feat: Refactor scheduler types and repository imports for improved organization
